src/services/route/RealTimeOptimizer.ts
import { WeatherService } from '../weather/WeatherService';
import { MCPIntegrationService } from '../integration/MCPIntegrationService';
import { CarRouteOptimizer } from './CarRouteOptimizer';
import { 
  ActivityType,
  ActivityMetrics,
  ActivitySegment,
  MultiSegmentActivity,
  DifficultyLevel,
  BikeMetrics,
  SkiMetrics,
  RunMetrics,
  WalkMetrics,
  CarMetrics
} from '@/types/activity/types';
import { 
  RoutePreferences,
  OptimizationResult,
  TerrainConditions
} from '@/types/route/types';
import { WeatherConditions } from '@/types/weather/types';
import { GeoPoint } from '@/types/geo';
import { RoutePreferences, RouteSegment, OptimizationType } from '@/types/route';
import { TerrainAnalysisService } from '../terrain/TerrainAnalysisService';
import { TrafficService } from '../traffic/TrafficService';

export class RealTimeOptimizer {
  private carOptimizer: CarRouteOptimizer;

  constructor(
    private weatherService: WeatherService,
    private mcpService: MCPIntegrationService,
    private terrainService: TerrainAnalysisService,
    private trafficService: TrafficService
  ) {
    this.carOptimizer = new CarRouteOptimizer(weatherService, mcpService);
  }

  async optimizeMultiSegmentRoute(
    segments: ActivitySegment[],
    preferences: RoutePreferences
  ): Promise<MultiSegmentActivity> {
    const optimizedSegments: ActivitySegment[] = [];
    let totalDistance = 0;
    let totalDuration = 0;
    let totalElevationGain = 0;
    let totalElevationLoss = 0;
    let totalCalories = 0;
    const transitions: MultiSegmentActivity['transitions'] = [];

    for (let i = 0; i < segments.length; i++) {
      const segment = segments[i];
      const weather = await this.weatherService.getWeatherForLocation(segment.startPoint);
      const terrain = await this.mcpService.getTerrainConditions(segment.startPoint);
      
      const optimizedResult = await this.optimizeSegment(
        segment.startPoint,
        segment.endPoint,
        segment.type,
        preferences,
        weather,
        terrain
      );

      const optimizedSegment: ActivitySegment = {
        ...segment,
        metrics: this.calculateActivityMetrics(optimizedResult, segment.type),
        waypoints: optimizedResult.waypoints
      };

      optimizedSegments.push(optimizedSegment);
      
      // Calculate segment metrics
      totalDistance += optimizedResult.metrics.distance;
      totalDuration += optimizedResult.metrics.duration;
      totalElevationGain += optimizedResult.metrics.elevation.gain;
      totalElevationLoss += optimizedResult.metrics.elevation.loss;
      totalCalories += optimizedResult.metrics.calories;

      // Handle transitions between segments
      if (i < segments.length - 1) {
        const nextSegment = segments[i + 1];
        const transition = await this.calculateTransition(
          optimizedSegment,
          nextSegment
        );
        transitions.push(transition);
        totalDuration += transition.duration;
      }
    }

    return {
      id: Date.now().toString(),
      segments: optimizedSegments,
      totalMetrics: {
        distance: totalDistance,
        duration: totalDuration,
        elevation: {
          gain: totalElevationGain,
          loss: totalElevationLoss
        },
        calories: totalCalories
      },
      transitions
    };
  }

  private async optimizeSegment(
    startPoint: { lat: number; lng: number; elevation?: number },
    endPoint: { lat: number; lng: number; elevation?: number },
    activityType: ActivityType,
    preferences: RoutePreferences,
    weather?: WeatherConditions,
    terrain?: TerrainConditions
  ): Promise<OptimizationResult> {
    // Activity-specific optimization logic
    switch (activityType) {
      case ActivityType.BIKE:
        return this.optimizeBikeRoute(startPoint, endPoint, preferences, weather, terrain);
      case ActivityType.SKI:
        return this.optimizeSkiRoute(startPoint, endPoint, preferences, weather, terrain);
      case ActivityType.CAR:
        return this.carOptimizer.optimizeRoute(startPoint, endPoint, preferences, weather, terrain);
      case ActivityType.RUN:
      case ActivityType.WALK:
        return this.optimizePedestrianRoute(startPoint, endPoint, preferences, weather, terrain);
      default:
        return this.optimizeDefaultRoute(startPoint, endPoint, preferences, weather, terrain);
    }
  }

  private calculateActivityMetrics(
    result: OptimizationResult,
    activityType: ActivityType
  ): ActivityMetrics {
    const baseMetrics = {
      speed: result.metrics.speed,
      distance: result.metrics.distance,
      duration: result.metrics.duration,
      elevation: result.metrics.elevation,
      calories: result.metrics.calories,
      pace: result.metrics.duration / result.metrics.distance,
      timestamp: new Date().toISOString(),
      weather: result.weather,
      terrain: result.terrain
    };

    switch (activityType) {
      case ActivityType.BIKE:
        return {
          ...baseMetrics,
          type: 'BIKE',
          cadence: 0, // To be calculated
          bikeLaneCoverage: result.metrics.bikeLaneCoverage || 0,
          trafficDensity: result.metrics.trafficDensity
        } as BikeMetrics & { type: 'BIKE' };

      case ActivityType.SKI:
        return {
          ...baseMetrics,
          type: 'SKI',
          snowDepth: result.metrics.snowDepth || 0,
          snowQuality: result.metrics.snowQuality || 'unknown',
          trailCondition: result.metrics.trailCondition || 'unknown',
          groomed: result.metrics.groomed || false,
          avalancheRisk: result.metrics.avalancheRisk
        } as SkiMetrics & { type: 'SKI' };

      case ActivityType.RUN:
        return {
          ...baseMetrics,
          type: 'RUN',
          cadence: 0, // To be calculated
          strideLength: 0, // To be calculated
        } as RunMetrics & { type: 'RUN' };

      case ActivityType.WALK:
        return {
          ...baseMetrics,
          type: 'WALK',
          stepCount: 0, // To be calculated
          terrainType: result.terrain?.type || 'unknown',
        } as WalkMetrics & { type: 'WALK' };

      case ActivityType.CAR:
        return {
          ...baseMetrics,
          type: 'CAR',
          fuelEfficiency: result.metrics.fuelEfficiency,
          trafficDelay: result.metrics.trafficDelay || 0,
          stopCount: result.metrics.stopCount || 0,
          alternativeRoutes: result.metrics.alternativeRoutes || 0
        } as CarMetrics & { type: 'CAR' };

      default:
        return {
          ...baseMetrics,
          type: 'BASE'
        };
    }
  }

  private async calculateTransition(
    fromSegment: ActivitySegment,
    toSegment: ActivitySegment
  ): Promise<MultiSegmentActivity['transitions'][0]> {
    const transitionPoint = fromSegment.endPoint;
    const estimatedDuration = await this.estimateTransitionDuration(
      fromSegment.type,
      toSegment.type,
      transitionPoint
    );

    return {
      fromType: fromSegment.type,
      toType: toSegment.type,
      point: transitionPoint,
      duration: estimatedDuration
    };
  }

  private async estimateTransitionDuration(
    fromType: ActivityType,
    toType: ActivityType,
    location: { lat: number; lng: number; elevation?: number }
  ): Promise<number> {
    // Base transition time in seconds
    let baseTime = 300; // 5 minutes default

    // Adjust based on activity types
    if (fromType === ActivityType.BIKE || toType === ActivityType.BIKE) {
      baseTime += 180; // Additional 3 minutes for bike preparation
    }
    if (fromType === ActivityType.SKI || toType === ActivityType.SKI) {
      baseTime += 600; // Additional 10 minutes for ski equipment
    }

    // Adjust for weather conditions
    const weather = await this.weatherService.getWeatherForLocation(location);
    if (weather?.conditions?.includes('rain') || weather?.conditions?.includes('snow')) {
      baseTime *= 1.5; // 50% longer in bad weather
    }

    return baseTime;
  }

  // Activity-specific optimization methods to be implemented
  private async optimizeBikeRoute(
    startPoint: { lat: number; lng: number; elevation?: number },
    endPoint: { lat: number; lng: number; elevation?: number },
    preferences: RoutePreferences,
    weather?: WeatherConditions,
    terrain?: TerrainConditions
  ): Promise<OptimizationResult> {
    try {
      const result: OptimizationResult = {
        path: [],
        metrics: {
          distance: 0,
          duration: 0,
          elevation: { gain: 0, loss: 0, profile: [] },
          safety: 1.0,
          weatherImpact: weather ? this.calculateWeatherImpact(weather, 'BIKE') : null,
          terrainDifficulty: terrain ? terrain.difficulty : 'unknown',
          surfaceType: terrain ? terrain.surface : 'unknown'
        },
        warnings: []
      };

      // Calculate optimal bike path considering elevation
      const elevationProfile = await this.calculateElevationProfile(startPoint, endPoint);
      result.metrics.elevation = elevationProfile;

      // Adjust for bike-specific preferences
      if (preferences.avoidHighways) {
        result.path = await this.calculateBikePathAvoidingHighways(startPoint, endPoint);
      } else {
        result.path = await this.calculateDefaultBikePath(startPoint, endPoint);
      }

      // Calculate safety score
      result.metrics.safety = this.calculateBikeSafetyScore(weather, terrain);

      return result;
    } catch (error) {
      console.error('Error optimizing bike route:', error);
      throw new Error('Failed to optimize bike route');
    }
  }

  private async optimizeSkiRoute(
    startPoint: { lat: number; lng: number; elevation?: number },
    endPoint: { lat: number; lng: number; elevation?: number },
    preferences: RoutePreferences,
    weather?: WeatherConditions,
    terrain?: TerrainConditions
  ): Promise<OptimizationResult> {
    try {
      const result: OptimizationResult = {
        path: [],
        metrics: {
          distance: 0,
          duration: 0,
          elevation: { gain: 0, loss: 0, profile: [] },
          safety: 1.0,
          weatherImpact: weather ? this.calculateWeatherImpact(weather, 'SKI') : null,
          terrainDifficulty: terrain ? terrain.difficulty : 'unknown',
          surfaceType: terrain ? terrain.surface : 'unknown'
        },
        warnings: []
      };

      // Check snow conditions
      if (weather && !this.isSnowConditionsSuitable(weather)) {
        result.warnings.push('poor_snow_conditions');
      }

      // Calculate optimal ski path considering slope
      const elevationProfile = await this.calculateElevationProfile(startPoint, endPoint);
      result.metrics.elevation = elevationProfile;

      // Adjust for ski-specific terrain
      if (terrain) {
        result.path = await this.calculateSkiPathForTerrain(startPoint, endPoint, terrain);
      } else {
        result.path = await this.calculateDefaultSkiPath(startPoint, endPoint);
      }

      // Calculate safety score
      result.metrics.safety = this.calculateSkiSafetyScore(weather, terrain);

      return result;
    } catch (error) {
      console.error('Error optimizing ski route:', error);
      throw new Error('Failed to optimize ski route');
    }
  }

  private async optimizePedestrianRoute(
    startPoint: { lat: number; lng: number; elevation?: number },
    endPoint: { lat: number; lng: number; elevation?: number },
    preferences: RoutePreferences,
    weather?: WeatherConditions,
    terrain?: TerrainConditions
  ): Promise<OptimizationResult> {
    try {
      const result: OptimizationResult = {
        path: [],
        metrics: {
          distance: 0,
          duration: 0,
          elevation: { gain: 0, loss: 0, profile: [] },
          safety: 1.0,
          weatherImpact: weather ? this.calculateWeatherImpact(weather, 'WALK') : null,
          terrainDifficulty: terrain ? terrain.difficulty : 'unknown',
          surfaceType: terrain ? terrain.surface : 'unknown'
        },
        warnings: []
      };

      // Calculate optimal pedestrian path
      const elevationProfile = await this.calculateElevationProfile(startPoint, endPoint);
      result.metrics.elevation = elevationProfile;

      // Adjust for pedestrian preferences
      if (preferences.preferScenic) {
        result.path = await this.calculateScenicPedestrianPath(startPoint, endPoint);
      } else if (preferences.avoidTraffic) {
        result.path = await this.calculateLowTrafficPedestrianPath(startPoint, endPoint);
      } else {
        result.path = await this.calculateDefaultPedestrianPath(startPoint, endPoint);
      }

      // Calculate safety score
      result.metrics.safety = this.calculatePedestrianSafetyScore(weather, terrain);

      return result;
    } catch (error) {
      console.error('Error optimizing pedestrian route:', error);
      throw new Error('Failed to optimize pedestrian route');
    }
  }

  private calculateWeatherImpact(weather: WeatherConditions, activity: string): number {
    let impact = 0;
    
    switch (activity) {
      case 'BIKE':
        if (weather.conditions.includes('rain')) impact += 0.4;
        if (weather.conditions.includes('snow')) impact += 0.8;
        if (weather.windSpeed > 20) impact += 0.5;
        break;
      case 'SKI':
        if (!weather.conditions.includes('snow')) impact += 0.9;
        if (weather.temperature > 5) impact += 0.7;
        if (weather.visibility < 1000) impact += 0.6;
        break;
      case 'WALK':
        if (weather.conditions.includes('rain')) impact += 0.3;
        if (weather.conditions.includes('snow')) impact += 0.4;
        if (weather.windSpeed > 30) impact += 0.3;
        break;
    }

    return Math.min(1, impact);
  }

  private calculateBikeSafetyScore(weather?: WeatherConditions, terrain?: TerrainConditions): number {
    let safety = 1.0;
    
    if (weather) {
      if (weather.conditions.includes('rain')) safety -= 0.3;
      if (weather.conditions.includes('snow')) safety -= 0.6;
      if (weather.windSpeed > 25) safety -= 0.4;
    }
    
    if (terrain) {
      if (terrain.surface === 'gravel') safety -= 0.2;
      if (terrain.surface === 'dirt') safety -= 0.3;
      if (terrain.surface === 'ice') safety -= 0.8;
    }
    
    return Math.max(0.1, safety);
  }

  private calculateSkiSafetyScore(weather?: WeatherConditions, terrain?: TerrainConditions): number {
    let safety = 1.0;
    
    if (weather) {
      if (!weather.conditions.includes('snow')) safety -= 0.5;
      if (weather.visibility < 1000) safety -= 0.4;
      if (weather.temperature > 5) safety -= 0.3;
    }
    
    if (terrain) {
      if (terrain.difficulty === 'expert') safety -= 0.4;
      if (terrain.hazards.includes('avalanche_risk')) safety -= 0.7;
    }
    
    return Math.max(0.1, safety);
  }

  private calculatePedestrianSafetyScore(weather?: WeatherConditions, terrain?: TerrainConditions): number {
    let safety = 1.0;
    
    if (weather) {
      if (weather.conditions.includes('ice')) safety -= 0.4;
      if (weather.visibility < 500) safety -= 0.3;
      if (weather.windSpeed > 40) safety -= 0.2;
    }
    
    if (terrain) {
      if (terrain.surface === 'ice') safety -= 0.5;
      if (terrain.difficulty === 'difficult') safety -= 0.3;
      terrain.hazards.forEach(() => safety -= 0.2);
    }
    
    return Math.max(0.1, safety);
  }

  private isSnowConditionsSuitable(weather: WeatherConditions): boolean {
    return weather.conditions.includes('snow') && 
           weather.temperature <= 5 && 
           weather.visibility >= 1000;
  }

  private async optimizeDefaultRoute(
    startPoint: { lat: number; lng: number; elevation?: number },
    endPoint: { lat: number; lng: number; elevation?: number },
    preferences: RoutePreferences,
    weather?: WeatherConditions,
    terrain?: TerrainConditions
  ): Promise<OptimizationResult> {
    // Implement default optimization
    throw new Error('Method not implemented.');
  }

  private async optimizeCarRoute(
    startPoint: { lat: number; lng: number; elevation?: number },
    endPoint: { lat: number; lng: number; elevation?: number },
    preferences: RoutePreferences,
    weather?: WeatherConditions,
    terrain?: TerrainConditions
  ): Promise<OptimizationResult> {
    // Get car-specific route data
    const trafficData = await this.mcpService.getTrafficData(startPoint, endPoint);
    const roadConditions = await this.mcpService.getRoadConditions(startPoint, endPoint);
    const restrictions = await this.mcpService.getRoadRestrictions(startPoint, endPoint);
    const elevationProfile = await this.mcpService.getElevationProfile(startPoint, endPoint);

    // Calculate optimal route considering car-specific factors
    const baseRoute = await this.mcpService.getBaseRoute(startPoint, endPoint);
    
    // Apply car-specific optimizations
    const optimizedPath = this.optimizeCarPath(
      baseRoute,
      trafficData,
      roadConditions,
      restrictions,
      elevationProfile,
      preferences
    );

    // Calculate metrics
    const distance = this.calculateRouteDistance(optimizedPath);
    const elevation = this.calculateElevationMetrics(optimizedPath, elevationProfile);
    const { duration, trafficDelay } = this.estimateCarDuration(
      distance,
      elevation,
      trafficData,
      weather,
      roadConditions
    );

    // Calculate car-specific metrics
    const { fuelEfficiency, stopCount } = this.calculateCarMetrics(
      optimizedPath,
      trafficData,
      elevation
    );
    const alternativeRoutes = await this.countAlternativeRoutes(
      startPoint,
      endPoint,
      preferences
    );

    return {
      path: optimizedPath,
      metrics: {
        distance,
        duration,
        elevation,
        speed: distance / duration,
        fuelEfficiency,
        trafficDelay,
        stopCount,
        alternativeRoutes
      },
      weather,
      terrain,
      waypoints: this.generateCarWaypoints(
        optimizedPath,
        trafficData,
        roadConditions
      ),
      alternativeRoutes: await this.generateCarAlternatives(
        startPoint,
        endPoint,
        preferences,
        weather,
        terrain
      ),
      warnings: this.generateCarWarnings(weather, roadConditions, trafficData)
    };
  }

  private optimizeCarPath(
    baseRoute: GeoPoint[],
    trafficData: any,
    roadConditions: any,
    restrictions: any,
    elevationProfile: any,
    preferences: RoutePreferences
  ): GeoPoint[] {
    let optimizedPath = [...baseRoute];

    // Avoid traffic congestion
    if (preferences.avoidTraffic) {
      optimizedPath = this.avoidTrafficCongestion(optimizedPath, trafficData);
    }

    // Consider road conditions
    optimizedPath = this.optimizeForRoadConditions(
      optimizedPath,
      roadConditions,
      preferences
    );

    // Handle road restrictions
    optimizedPath = this.applyRoadRestrictions(optimizedPath, restrictions);

    // Optimize for fuel efficiency if requested
    if (preferences.optimize === 'FUEL_EFFICIENCY') {
      optimizedPath = this.optimizeForFuelEfficiency(
        optimizedPath,
        elevationProfile,
        trafficData
      );
    }

    return optimizedPath;
  }

  private estimateCarDuration(
    distance: number,
    elevation: { gain: number; loss: number },
    trafficData: any,
    weather?: WeatherConditions,
    roadConditions?: any
  ): { duration: number; trafficDelay: number } {
    // Base duration calculation (assuming average speed of 50 km/h in urban areas)
    const baseSpeed = 50;
    let duration = distance / (baseSpeed * 1000 / 3600); // Convert to seconds
    let trafficDelay = 0;

    // Traffic adjustments
    if (trafficData) {
      const congestionFactor = this.calculateCongestionFactor(trafficData);
      const trafficDuration = duration * congestionFactor;
      trafficDelay = trafficDuration - duration;
      duration = trafficDuration;
    }

    // Weather adjustments
    if (weather) {
      if (weather.conditions.includes('rain')) {
        duration *= 1.2; // 20% longer in rain
      }
      if (weather.conditions.includes('snow')) {
        duration *= 1.5; // 50% longer in snow
      }
      if (weather.visibility < 1000) {
        duration *= 1.3; // 30% longer in poor visibility
      }
    }

    // Road condition adjustments
    if (roadConditions) {
      switch (roadConditions.overall) {
        case 'poor':
          duration *= 1.3;
          break;
        case 'fair':
          duration *= 1.1;
          break;
        case 'construction':
          duration *= 1.4;
          break;
      }
    }

    return {
      duration: Math.round(duration),
      trafficDelay: Math.round(trafficDelay)
    };
  }

  private calculateCarMetrics(
    path: GeoPoint[],
    trafficData: any,
    elevation: { gain: number; loss: number }
  ): { fuelEfficiency: number; stopCount: number } {
    // Calculate number of stops (traffic lights, stop signs, congestion)
    const stopCount = this.calculateStopCount(path, trafficData);

    // Calculate fuel efficiency (km/L)
    // Base efficiency of 10 km/L
    let fuelEfficiency = 10;

    // Adjust for elevation changes
    const elevationImpact = 1 - (elevation.gain / 1000) * 0.1;
    fuelEfficiency *= elevationImpact;

    // Adjust for traffic conditions
    const trafficImpact = 1 - (stopCount / 10) * 0.05;
    fuelEfficiency *= trafficImpact;

    return {
      fuelEfficiency: Math.round(fuelEfficiency * 10) / 10,
      stopCount
    };
  }

  private calculateStopCount(path: GeoPoint[], trafficData: any): number {
    let stops = 0;

    // Count traffic lights
    stops += trafficData.trafficLights.length;

    // Count stop signs
    stops += trafficData.stopSigns.length;

    // Add stops due to congestion
    const congestionStops = Math.floor(
      trafficData.congestionLevel * trafficData.segments.length * 0.2
    );
    stops += congestionStops;

    return stops;
  }

  private async countAlternativeRoutes(
    startPoint: { lat: number; lng: number; elevation?: number },
    endPoint: { lat: number; lng: number; elevation?: number },
    preferences: RoutePreferences
  ): Promise<number> {
    const alternatives = await this.mcpService.getAlternativeRoutes(
      startPoint,
      endPoint,
      preferences
    );
    return alternatives.length;
  }

  private generateCarWaypoints(
    path: GeoPoint[],
    trafficData: any,
    roadConditions: any
  ): Array<{ lat: number; lng: number; elevation?: number; type: string }> {
    const waypoints = [];

    // Add traffic incident points
    for (const incident of trafficData.incidents) {
      waypoints.push({
        ...incident.location,
        type: 'traffic_incident'
      });
    }

    // Add road condition warning points
    for (const warning of roadConditions.warnings) {
      waypoints.push({
        ...warning.location,
        type: 'road_condition'
      });
    }

    // Add major intersections
    for (const intersection of trafficData.majorIntersections) {
      waypoints.push({
        ...intersection.location,
        type: 'major_intersection'
      });
    }

    return waypoints;
  }

  private async generateCarAlternatives(
    startPoint: { lat: number; lng: number; elevation?: number },
    endPoint: { lat: number; lng: number; elevation?: number },
    preferences: RoutePreferences,
    weather?: WeatherConditions,
    terrain?: TerrainConditions
  ): Promise<Array<{ path: GeoPoint[]; metrics: any }>> {
    const alternatives = [];

    // Generate fastest route alternative
    const fastestRoute = await this.mcpService.getFastestCarRoute(
      startPoint,
      endPoint,
      weather
    );
    alternatives.push(fastestRoute);

    // Generate fuel-efficient route alternative
    const ecoRoute = await this.mcpService.getFuelEfficientRoute(
      startPoint,
      endPoint,
      weather
    );
    alternatives.push(ecoRoute);

    // Generate scenic route alternative if preferred
    if (preferences.preferScenic) {
      const scenicRoute = await this.mcpService.getScenicCarRoute(
        startPoint,
        endPoint
      );
      alternatives.push(scenicRoute);
    }

    return alternatives;
  }

  private generateCarWarnings(
    weather?: WeatherConditions,
    roadConditions?: any,
    trafficData?: any
  ): string[] {
    const warnings = [];

    // Weather warnings
    if (weather) {
      if (weather.conditions.includes('snow')) {
        warnings.push('Snowy conditions - reduce speed');
      }
      if (weather.conditions.includes('rain')) {
        warnings.push('Wet roads - use caution');
      }
      if (weather.visibility < 1000) {
        warnings.push('Poor visibility conditions');
      }
      if (weather.windSpeed > 50) {
        warnings.push('Strong winds - high-profile vehicles use caution');
      }
    }

    // Road condition warnings
    if (roadConditions) {
      if (roadConditions.overall === 'poor') {
        warnings.push('Poor road conditions ahead');
      }
      if (roadConditions.construction) {
        warnings.push('Construction zones on route');
      }
      if (roadConditions.hazards.length > 0) {
        warnings.push('Road hazards reported');
      }
    }

    // Traffic warnings
    if (trafficData) {
      if (trafficData.congestionLevel > 0.7) {
        warnings.push('Heavy traffic conditions');
      }
      if (trafficData.incidents.length > 0) {
        warnings.push('Traffic incidents reported');
      }
      if (trafficData.slowdowns > 2) {
        warnings.push('Multiple slowdown areas');
      }
    }

    return warnings;
  }

  private avoidTrafficCongestion(
    path: GeoPoint[],
    trafficData: any
  ): GeoPoint[] {
    // Find segments with high congestion
    const congestionPoints = trafficData.segments
      .filter(segment => segment.congestionLevel > 0.7)
      .map(segment => ({
        start: segment.start,
        end: segment.end
      }));

    if (congestionPoints.length === 0) {
      return path;
    }

    // Reroute around congested segments
    return this.findAlternativeRoute(path, congestionPoints);
  }

  private optimizeForRoadConditions(
    path: GeoPoint[],
    roadConditions: any,
    preferences: RoutePreferences
  ): GeoPoint[] {
    // Find segments with poor conditions
    const poorConditionSegments = roadConditions.segments
      .filter(segment => segment.condition === 'poor')
      .map(segment => ({
        start: segment.start,
        end: segment.end
      }));

    if (poorConditionSegments.length === 0) {
      return path;
    }

    // Reroute around poor condition segments if possible
    return this.findAlternativeRoute(path, poorConditionSegments);
  }

  private applyRoadRestrictions(
    path: GeoPoint[],
    restrictions: any
  ): GeoPoint[] {
    // Find segments with restrictions
    const restrictedSegments = restrictions.segments
      .filter(segment => segment.restricted)
      .map(segment => ({
        start: segment.start,
        end: segment.end
      }));

    if (restrictedSegments.length === 0) {
      return path;
    }

    // Reroute around restricted segments
    return this.findAlternativeRoute(path, restrictedSegments);
  }

  private optimizeForFuelEfficiency(
    path: GeoPoint[],
    elevationProfile: any,
    trafficData: any
  ): GeoPoint[] {
    // Find segments with high fuel consumption
    const inefficientSegments = this.findInefficientSegments(
      path,
      elevationProfile,
      trafficData
    );

    if (inefficientSegments.length === 0) {
      return path;
    }

    // Reroute for better fuel efficiency
    return this.findEfficientRoute(path, inefficientSegments);
  }

  private findInefficientSegments(
    path: GeoPoint[],
    elevationProfile: any,
    trafficData: any
  ): Array<{ start: GeoPoint; end: GeoPoint }> {
    const inefficientSegments = [];

    for (let i = 0; i < path.length - 1; i++) {
      const segment = {
        start: path[i],
        end: path[i + 1]
      };

      // Check elevation change
      const elevationChange = Math.abs(
        elevationProfile[i + 1] - elevationProfile[i]
      );
      const distance = this.calculateHaversineDistance(
        segment.start,
        segment.end
      );
      const grade = elevationChange / distance;

      // Check traffic congestion
      const congestion = this.getSegmentCongestion(segment, trafficData);

      // If either grade is too steep or congestion is too high
      if (grade > 0.1 || congestion > 0.7) {
        inefficientSegments.push(segment);
      }
    }

    return inefficientSegments;
  }

  private getSegmentCongestion(
    segment: { start: GeoPoint; end: GeoPoint },
    trafficData: any
  ): number {
    // Find the traffic data for this segment
    const trafficSegment = trafficData.segments.find(
      (s: any) =>
        this.calculateHaversineDistance(s.start, segment.start) < 100 &&
        this.calculateHaversineDistance(s.end, segment.end) < 100
    );

    return trafficSegment ? trafficSegment.congestionLevel : 0;
  }

  private findAlternativeRoute(
    path: GeoPoint[],
    segmentsToAvoid: Array<{ start: GeoPoint; end: GeoPoint }>
  ): GeoPoint[] {
    // This is a placeholder implementation
    // In a real implementation, this would use a pathfinding algorithm
    // to find an alternative route avoiding the specified segments
    return path;
  }

  private findEfficientRoute(
    path: GeoPoint[],
    inefficientSegments: Array<{ start: GeoPoint; end: GeoPoint }>
  ): GeoPoint[] {
    // This is a placeholder implementation
    // In a real implementation, this would use a pathfinding algorithm
    // to find a more fuel-efficient route
    return path;
  }

  private calculateCongestionFactor(trafficData: any): number {
    // Calculate weighted average of congestion levels
    let totalWeight = 0;
    let weightedCongestion = 0;

    for (const segment of trafficData.segments) {
      const weight = segment.length;
      weightedCongestion += segment.congestionLevel * weight;
      totalWeight += weight;
    }

    const avgCongestion = weightedCongestion / totalWeight;
    
    // Convert to factor (1.0 = no delay, 2.0 = twice as long, etc.)
    return 1 + avgCongestion;
  }

  async optimizeRoute(
    start: GeoPoint,
    end: GeoPoint,
    preferences: RoutePreferences,
    currentConditions?: {
      weather?: WeatherConditions;
      terrain?: TerrainConditions;
    }
  ) {
    const weather = currentConditions?.weather || await this.weatherService.getWeatherForLocation(start);
    const terrain = currentConditions?.terrain || await this.terrainService.getTerrainConditions(start);
    
    // Get initial route from MCP
    let route = await this.mcpService.getRoute(start, end, preferences);
    
    // Apply dynamic optimizations
    route = await this.applyDynamicOptimizations(route, preferences, weather, terrain);
    
    // Predict and avoid congestion
    route = await this.avoidPredictedCongestion(route, preferences);
    
    // Generate and rank alternatives
    const alternatives = await this.generateAlternativeRoutes(route, preferences);
    
    return {
      primary: route,
      alternatives: this.rankAlternativeRoutes(alternatives, preferences)
    };
  }

  private async applyDynamicOptimizations(
    route: RouteSegment[],
    preferences: RoutePreferences,
    weather: WeatherConditions,
    terrain: TerrainConditions
  ): Promise<RouteSegment[]> {
    // Apply weather-based optimizations
    if (weather.severity > 0.5) {
      route = await this.optimizeForWeather(route, weather, preferences);
    }

    // Apply terrain-based optimizations
    if (terrain.difficulty !== 'easy') {
      route = await this.optimizeForTerrain(route, terrain, preferences);
    }

    // Apply time-based optimizations
    route = await this.optimizeForTimeOfDay(route, preferences);

    return route;
  }

  private async avoidPredictedCongestion(
    route: RouteSegment[],
    preferences: RoutePreferences
  ): Promise<RouteSegment[]> {
    const congestionPredictions = await Promise.all(
      route.map(segment => this.trafficService.predictCongestion(segment.start, segment.end))
    );

    // Reroute segments with high predicted congestion
    const optimizedSegments = await Promise.all(
      route.map(async (segment, index) => {
        if (congestionPredictions[index] > 0.7) {
          return this.findAlternativeSegment(segment, preferences);
        }
        return segment;
      })
    );

    return optimizedSegments;
  }

  private async generateAlternativeRoutes(
    primaryRoute: RouteSegment[],
    preferences: RoutePreferences
  ): Promise<RouteSegment[][]> {
    const start = primaryRoute[0].start;
    const end = primaryRoute[primaryRoute.length - 1].end;

    // Generate alternatives with different optimization criteria
    const alternatives = await Promise.all([
      this.mcpService.getRoute(start, end, { ...preferences, optimize: 'TIME' }),
      this.mcpService.getRoute(start, end, { ...preferences, optimize: 'DISTANCE' }),
      this.mcpService.getRoute(start, end, { ...preferences, optimize: 'SAFETY' })
    ]);

    return alternatives;
  }

  private rankAlternativeRoutes(
    alternatives: RouteSegment[][],
    preferences: RoutePreferences
  ): RouteSegment[][] {
    return alternatives.sort((a, b) => {
      const scoreA = this.calculateRouteScore(a, preferences);
      const scoreB = this.calculateRouteScore(b, preferences);
      return scoreB - scoreA;
    });
  }

  private calculateRouteScore(route: RouteSegment[], preferences: RoutePreferences): number {
    const weights = preferences.weights;
    let score = 0;

    // Calculate weighted score based on metrics
    route.forEach(segment => {
      score += segment.metrics.distance * weights.distance;
      score += segment.metrics.duration * weights.duration;
      score += segment.metrics.safety * weights.safety;
      score += this.calculateComfortScore(segment) * weights.comfort;
    });

    return score;
  }

  private calculateComfortScore(segment: RouteSegment): number {
    // Implement comfort scoring based on surface type, elevation changes, etc.
    let score = 1.0;

    if (segment.conditions) {
      // Reduce score for difficult terrain
      if (segment.conditions.difficulty !== 'easy') {
        score *= 0.8;
      }

      // Reduce score for hazards
      score *= Math.max(0.5, 1 - (segment.conditions.hazards.length * 0.1));

      // Adjust for surface quality
      if (segment.conditions.quality) {
        score *= segment.conditions.quality.stability;
      }
    }

    return score;
  }

  private async findAlternativeSegment(
    segment: RouteSegment,
    preferences: RoutePreferences
  ): Promise<RouteSegment> {
    // Get alternative segments avoiding congested areas
    const alternatives = await this.mcpService.getAlternativeSegments(
      segment.start,
      segment.end,
      {
        ...preferences,
        avoidTraffic: true
      }
    );

    // Return the best alternative based on current conditions
    return alternatives[0] || segment;
  }

  private async optimizeForWeather(
    route: RouteSegment[],
    weather: WeatherConditions,
    preferences: RoutePreferences
  ): Promise<RouteSegment[]> {
    // Implement weather-based optimizations
    return route.map(segment => ({
      ...segment,
      metrics: {
        ...segment.metrics,
        safety: this.adjustSafetyForWeather(segment.metrics.safety, weather)
      }
    }));
  }

  private async optimizeForTerrain(
    route: RouteSegment[],
    terrain: TerrainConditions,
    preferences: RoutePreferences
  ): Promise<RouteSegment[]> {
    // Implement terrain-based optimizations
    return route.map(segment => ({
      ...segment,
      metrics: {
        ...segment.metrics,
        safety: this.adjustSafetyForTerrain(segment.metrics.safety, terrain)
      }
    }));
  }

  private async optimizeForTimeOfDay(
    route: RouteSegment[],
    preferences: RoutePreferences
  ): Promise<RouteSegment[]> {
    const currentHour = new Date().getHours();
    const isRushHour = (currentHour >= 7 && currentHour <= 9) || (currentHour >= 16 && currentHour <= 18);

    if (isRushHour && preferences.avoidTraffic) {
      return this.avoidPredictedCongestion(route, preferences);
    }

    return route;
  }

  private adjustSafetyForWeather(safety: number, weather: WeatherConditions): number {
    let adjustment = 1.0;

    // Reduce safety score based on weather conditions
    if (weather.conditions.includes('rain')) adjustment *= 0.8;
    if (weather.conditions.includes('snow')) adjustment *= 0.6;
    if (weather.conditions.includes('ice')) adjustment *= 0.4;
    if (weather.visibility < 5000) adjustment *= 0.7;

    return safety * adjustment;
  }

  private adjustSafetyForTerrain(safety: number, terrain: TerrainConditions): number {
    let adjustment = 1.0;

    // Reduce safety score based on terrain conditions
    if (terrain.difficulty === 'difficult') adjustment *= 0.7;
    if (terrain.difficulty === 'expert') adjustment *= 0.5;
    if (terrain.difficulty === 'extreme') adjustment *= 0.3;

    // Further reduce for hazards
    adjustment *= Math.max(0.3, 1 - (terrain.hazards.length * 0.15));

    return safety * adjustment;
  }
} 